Updating code to use a composition based approach
