+++ iothub/device/src/ClientFactory.cs
@@ -1,0 +1,485 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.Azure.Devices.Client.Edge;
+
+namespace Microsoft.Azure.Devices.Client
+{
+    using System;
+    using System.Text.RegularExpressions;
+    using Microsoft.Azure.Devices.Client.Extensions;
+    using Microsoft.Azure.Devices.Client.Transport;
+    using Microsoft.Azure.Devices.Client.Transport.Mqtt;
+
+    internal class ClientFactory
+    {
+        const string DeviceId = "DeviceId";
+        const string DeviceIdParameterPattern = @"(^\s*?|.*;\s*?)" + DeviceId + @"\s*?=.*";
+        const RegexOptions RegexOptions = System.Text.RegularExpressions.RegexOptions.Compiled | System.Text.RegularExpressions.RegexOptions.IgnoreCase;
+        static readonly Regex DeviceIdParameterRegex = new Regex(DeviceIdParameterPattern, RegexOptions);
+
+        /// <summary>
+        /// Create an Amqp InternalClient from individual parameters
+        /// </summary>
+        /// <param name="hostname">The fully-qualified DNS hostname of IoT Hub</param>
+        /// <param name="authenticationMethod">The authentication method that is used</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient Create(string hostname, IAuthenticationMethod authenticationMethod)
+        {
+            return Create(hostname, authenticationMethod, TransportType.Amqp);
+        }
+
+        /// <summary>
+        /// Create an Amqp InternalClient from individual parameters
+        /// </summary>
+        /// <param name="hostname">The fully-qualified DNS hostname of IoT Hub</param>
+        /// <param name="authenticationMethod">The authentication method that is used</param>
+        /// <param name="gatewayHostname">The fully-qualified DNS hostname of Gateway</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient Create(string hostname, string gatewayHostname, IAuthenticationMethod authenticationMethod)
+        {
+            return Create(hostname, gatewayHostname, authenticationMethod, TransportType.Amqp);
+        }
+
+        /// <summary>
+        /// Create a InternalClient from individual parameters
+        /// </summary>
+        /// <param name="hostname">The fully-qualified DNS hostname of IoT Hub</param>
+        /// <param name="authenticationMethod">The authentication method that is used</param>
+        /// <param name="transportType">The transportType used (Http1 or Amqp)</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient Create(string hostname, IAuthenticationMethod authenticationMethod, TransportType transportType)
+        {
+            return Create(hostname, null, authenticationMethod, transportType);
+        }
+
+        /// <summary>
+        /// Create a InternalClient from individual parameters
+        /// </summary>
+        /// <param name="hostname">The fully-qualified DNS hostname of IoT Hub</param>
+        /// <param name="gatewayHostname">The fully-qualified DNS hostname of Gateway</param>
+        /// <param name="authenticationMethod">The authentication method that is used</param>
+        /// <param name="transportType">The transportType used (Http1 or Amqp)</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient Create(string hostname, string gatewayHostname, IAuthenticationMethod authenticationMethod, TransportType transportType)
+        {
+            if (hostname == null)
+            {
+                throw new ArgumentNullException(nameof(hostname));
+            }
+
+            if (authenticationMethod == null)
+            {
+                throw new ArgumentNullException(nameof(authenticationMethod));
+            }
+
+            IotHubConnectionStringBuilder connectionStringBuilder = IotHubConnectionStringBuilder.Create(hostname, gatewayHostname, authenticationMethod);
+
+#if !NETMF
+            if (authenticationMethod is DeviceAuthenticationWithX509Certificate)
+            {
+                if (connectionStringBuilder.Certificate == null)
+                {
+                    throw new ArgumentException("certificate must be present in DeviceAuthenticationWithX509Certificate");
+                }
+
+                if (!connectionStringBuilder.Certificate.HasPrivateKey)
+                {
+                    throw new ArgumentException("certificate in DeviceAuthenticationWithX509Certificate must have a private key");
+                }
+
+                InternalClient dc = CreateFromConnectionString(connectionStringBuilder.ToString(), PopulateCertificateInTransportSettings(connectionStringBuilder, transportType));
+                dc.Certificate = connectionStringBuilder.Certificate;
+                return dc;
+            }
+#endif
+
+            return CreateFromConnectionString(connectionStringBuilder.ToString(), authenticationMethod, transportType, null);
+        }
+
+        /// <summary>
+        /// Create a InternalClient from individual parameters
+        /// </summary>
+        /// <param name="hostname">The fully-qualified DNS hostname of IoT Hub</param>
+        /// <param name="authenticationMethod">The authentication method that is used</param>
+        /// <param name="transportSettings">Prioritized list of transportTypes and their settings</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient Create(string hostname, IAuthenticationMethod authenticationMethod,
+            ITransportSettings[] transportSettings)
+        {
+            return Create(hostname, null, authenticationMethod, transportSettings);
+        }
+
+        /// <summary>
+        /// Create a InternalClient from individual parameters
+        /// </summary>
+        /// <param name="hostname">The fully-qualified DNS hostname of IoT Hub</param>
+        /// <param name="gatewayHostname">The fully-qualified DNS hostname of Gateway</param>
+        /// <param name="authenticationMethod">The authentication method that is used</param>
+        /// <param name="transportSettings">Prioritized list of transportTypes and their settings</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient Create(string hostname, string gatewayHostname, IAuthenticationMethod authenticationMethod,
+            ITransportSettings[] transportSettings)
+        {
+            if (hostname == null)
+            {
+                throw new ArgumentNullException("hostname");
+            }
+
+            if (authenticationMethod == null)
+            {
+                throw new ArgumentNullException("authenticationMethod");
+            }
+
+            var connectionStringBuilder = IotHubConnectionStringBuilder.Create(hostname, gatewayHostname, authenticationMethod);
+#if !NETMF
+            if (authenticationMethod is DeviceAuthenticationWithX509Certificate)
+            {
+                if (connectionStringBuilder.Certificate == null)
+                {
+                    throw new ArgumentException("certificate must be present in DeviceAuthenticationWithX509Certificate");
+                }
+
+                if (!connectionStringBuilder.Certificate.HasPrivateKey)
+                {
+                    throw new ArgumentException("certificate in DeviceAuthenticationWithX509Certificate must have a private key");
+                }
+
+                InternalClient dc = CreateFromConnectionString(connectionStringBuilder.ToString(), PopulateCertificateInTransportSettings(connectionStringBuilder, transportSettings));
+                dc.Certificate = connectionStringBuilder.Certificate;
+                return dc;
+            }
+#endif
+            return CreateFromConnectionString(connectionStringBuilder.ToString(), authenticationMethod, transportSettings, null);
+        }
+
+        /// <summary>
+        /// Create a InternalClient using Amqp transport from the specified connection string
+        /// </summary>
+        /// <param name="connectionString">Connection string for the IoT hub (including DeviceId)</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient CreateFromConnectionString(string connectionString)
+        {
+            return CreateFromConnectionString(connectionString, TransportType.Amqp);
+        }
+
+        /// <summary>
+        /// Create a InternalClient using Amqp transport from the specified connection string
+        /// </summary>
+        /// <param name="connectionString">IoT Hub-Scope Connection string for the IoT hub (without DeviceId)</param>
+        /// <param name="deviceId">Id of the Device</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient CreateFromConnectionString(string connectionString, string deviceId)
+        {
+            return CreateFromConnectionString(connectionString, deviceId, TransportType.Amqp);
+        }
+
+        /// <summary>
+        /// Create InternalClient from the specified connection string using the specified transport type
+        /// </summary>
+        /// <param name="connectionString">Connection string for the IoT hub (including DeviceId)</param>
+        /// <param name="transportType">Specifies whether Amqp or Http transport is used</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient CreateFromConnectionString(string connectionString, TransportType transportType)
+        {
+            if (string.IsNullOrEmpty(connectionString))
+            {
+                throw new ArgumentNullException(nameof(connectionString));
+            }
+
+            return CreateFromConnectionString(connectionString, null, transportType, null);
+        }
+
+        /// <summary>
+        /// Create InternalClient from the specified connection string using the specified transport type
+        /// </summary>
+        /// <param name="connectionString">IoT Hub-Scope Connection string for the IoT hub (without DeviceId)</param>
+        /// <param name="deviceId">Id of the device</param>
+        /// <param name="transportType">The transportType used (Http1 or Amqp)</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient CreateFromConnectionString(string connectionString, string deviceId, TransportType transportType)
+        {
+            if (connectionString == null)
+            {
+                throw new ArgumentNullException(nameof(connectionString));
+            }
+
+            if (deviceId == null)
+            {
+                throw new ArgumentNullException(nameof(deviceId));
+            }
+
+            if (DeviceIdParameterRegex.IsMatch(connectionString))
+            {
+                throw new ArgumentException("Connection string must not contain DeviceId keyvalue parameter", nameof(connectionString));
+            }
+
+            return CreateFromConnectionString(connectionString + ";" + DeviceId + "=" + deviceId, transportType);
+        }
+
+        /// <summary>
+        /// Create InternalClient from the specified connection string using a prioritized list of transports
+        /// </summary>
+        /// <param name="connectionString">Connection string for the IoT hub (with DeviceId)</param>
+        /// <param name="transportSettings">Prioritized list of transports and their settings</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient CreateFromConnectionString(string connectionString,
+            ITransportSettings[] transportSettings)
+        {
+            return CreateFromConnectionString(connectionString, null, transportSettings, null);
+        }
+
+
+        /// <summary>
+        /// Create InternalClient from the specified connection string using the prioritized list of transports
+        /// </summary>
+        /// <param name="connectionString">Connection string for the IoT hub (without DeviceId)</param>
+        /// <param name="deviceId">Id of the device</param>
+        /// <param name="transportSettings">Prioritized list of transportTypes and their settings</param>
+        /// <returns>InternalClient</returns>
+        public static InternalClient CreateFromConnectionString(string connectionString, string deviceId,
+            ITransportSettings[] transportSettings)
+        {
+            if (connectionString == null)
+            {
+                throw new ArgumentNullException(nameof(connectionString));
+            }
+
+            if (deviceId == null)
+            {
+                throw new ArgumentNullException(nameof(deviceId));
+            }
+
+            if (DeviceIdParameterRegex.IsMatch(connectionString))
+            {
+                throw new ArgumentException("Connection string must not contain DeviceId keyvalue parameter", nameof(connectionString));
+            }
+
+            return CreateFromConnectionString(connectionString + ";" + DeviceId + "=" + deviceId, transportSettings);
+        }
+
+        internal static InternalClient CreateFromConnectionString(
+            string connectionString,
+            IAuthenticationMethod authenticationMethod,
+            TransportType transportType,
+            IDeviceClientPipelineBuilder pipelineBuilder)
+        {
+            ITransportSettings[] transportSettings = GetTransportSettings(transportType);
+            return CreateFromConnectionString(connectionString, authenticationMethod, transportSettings,
+                pipelineBuilder);
+        }
+
+        internal static ITransportSettings[] GetTransportSettings(TransportType transportType)
+        {
+            switch (transportType)
+            {
+                case TransportType.Amqp:
+                    return new ITransportSettings[]
+                    {
+                        new AmqpTransportSettings(TransportType.Amqp_Tcp_Only),
+                        new AmqpTransportSettings(TransportType.Amqp_WebSocket_Only)
+                    };
+
+                case TransportType.Mqtt:
+                    return new ITransportSettings[]
+                    {
+                        new MqttTransportSettings(TransportType.Mqtt_Tcp_Only),
+                        new MqttTransportSettings(TransportType.Mqtt_WebSocket_Only)
+                    };
+
+                case TransportType.Amqp_WebSocket_Only:
+                case TransportType.Amqp_Tcp_Only:
+                    return new ITransportSettings[]
+                    {
+                        new AmqpTransportSettings(transportType)
+                    };
+
+                case TransportType.Mqtt_WebSocket_Only:
+                case TransportType.Mqtt_Tcp_Only:
+                    return new ITransportSettings[]
+                    {
+                        new MqttTransportSettings(transportType)
+                    };
+
+                case TransportType.Http1:
+                    return new ITransportSettings[] { new Http1TransportSettings() };
+
+                default:
+                    throw new InvalidOperationException("Unsupported Transport Type {0}".FormatInvariant(transportType));
+            }
+        }
+
+        internal static InternalClient CreateFromConnectionString(
+            string connectionString,
+            IAuthenticationMethod authenticationMethod,
+            ITransportSettings[] transportSettings,
+            IDeviceClientPipelineBuilder pipelineBuilder)
+        {
+            if (connectionString == null)
+            {
+                throw new ArgumentNullException(nameof(connectionString));
+            }
+
+            if (transportSettings == null)
+            {
+                throw new ArgumentNullException(nameof(transportSettings));
+            }
+
+            if (transportSettings.Length == 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(connectionString), "Must specify at least one TransportSettings instance");
+            }
+
+            var builder = IotHubConnectionStringBuilder.CreateWithIAuthenticationOverride(
+                connectionString,
+                authenticationMethod);
+
+            IotHubConnectionString iotHubConnectionString = builder.ToIotHubConnectionString();
+
+            foreach (ITransportSettings transportSetting in transportSettings)
+            {
+                switch (transportSetting.GetTransportType())
+                {
+                    case TransportType.Amqp_WebSocket_Only:
+                    case TransportType.Amqp_Tcp_Only:
+                        if (!(transportSetting is AmqpTransportSettings))
+                        {
+                            throw new InvalidOperationException("Unknown implementation of ITransportSettings type");
+                        }
+                        break;
+                    case TransportType.Http1:
+                        if (!(transportSetting is Http1TransportSettings))
+                        {
+                            throw new InvalidOperationException("Unknown implementation of ITransportSettings type");
+                        }
+                        break;
+                    case TransportType.Mqtt_WebSocket_Only:
+                    case TransportType.Mqtt_Tcp_Only:
+                        if (!(transportSetting is MqttTransportSettings))
+                        {
+                            throw new InvalidOperationException("Unknown implementation of ITransportSettings type");
+                        }
+                        break;
+                    default:
+                        throw new InvalidOperationException("Unsupported Transport Type {0}".FormatInvariant(transportSetting.GetTransportType()));
+                }
+            }
+
+            pipelineBuilder = pipelineBuilder ?? BuildPipeline();
+
+            // Defer concrete InternalClient creation to OpenAsync
+            return new InternalClient(iotHubConnectionString, transportSettings, pipelineBuilder);
+        }
+
+
+        static IDeviceClientPipelineBuilder BuildPipeline()
+        {
+            var transporthandlerFactory = new TransportHandlerFactory();
+            IDeviceClientPipelineBuilder pipelineBuilder = new DeviceClientPipelineBuilder()
+                .With(ctx => new GateKeeperDelegatingHandler(ctx))
+                .With(ctx => new RetryDelegatingHandler(ctx))
+                .With(ctx => new ErrorDelegatingHandler(ctx))
+                .With(ctx => new ProtocolRoutingDelegatingHandler(ctx))
+                .With(ctx => transporthandlerFactory.Create(ctx));
+            return pipelineBuilder;
+        }
+
+        static ITransportSettings[] PopulateCertificateInTransportSettings(IotHubConnectionStringBuilder connectionStringBuilder, TransportType transportType)
+        {
+            switch (transportType)
+            {
+                case TransportType.Amqp:
+                    return new ITransportSettings[]
+                    {
+                        new AmqpTransportSettings(TransportType.Amqp_Tcp_Only)
+                        {
+                            ClientCertificate = connectionStringBuilder.Certificate
+                        },
+                        new AmqpTransportSettings(TransportType.Amqp_WebSocket_Only)
+                        {
+                            ClientCertificate = connectionStringBuilder.Certificate
+                        }
+                    };
+                case TransportType.Amqp_Tcp_Only:
+                    return new ITransportSettings[]
+                    {
+                        new AmqpTransportSettings(TransportType.Amqp_Tcp_Only)
+                        {
+                            ClientCertificate = connectionStringBuilder.Certificate
+                        }
+                    };
+                case TransportType.Amqp_WebSocket_Only:
+                    return new ITransportSettings[]
+                    {
+                        new AmqpTransportSettings(TransportType.Amqp_WebSocket_Only)
+                        {
+                            ClientCertificate = connectionStringBuilder.Certificate
+                        }
+                    };
+                case TransportType.Http1:
+                    return new ITransportSettings[]
+                    {
+                        new Http1TransportSettings()
+                        {
+                            ClientCertificate = connectionStringBuilder.Certificate
+                        }
+                    };
+                case TransportType.Mqtt:
+                    return new ITransportSettings[]
+                    {
+                        new MqttTransportSettings(TransportType.Mqtt_Tcp_Only)
+                        {
+                            ClientCertificate = connectionStringBuilder.Certificate
+                        },
+                        new MqttTransportSettings(TransportType.Mqtt_WebSocket_Only)
+                        {
+                            ClientCertificate = connectionStringBuilder.Certificate
+                        }
+                    };
+                case TransportType.Mqtt_Tcp_Only:
+                    return new ITransportSettings[]
+                    {
+                        new MqttTransportSettings(TransportType.Mqtt_Tcp_Only)
+                        {
+                            ClientCertificate = connectionStringBuilder.Certificate
+                        }
+                    };
+                case TransportType.Mqtt_WebSocket_Only:
+                    return new ITransportSettings[]
+                    {
+                        new MqttTransportSettings(TransportType.Mqtt_WebSocket_Only)
+                        {
+                            ClientCertificate = connectionStringBuilder.Certificate
+                        }
+                    };
+                default:
+                    throw new InvalidOperationException("Unsupported Transport {0}".FormatInvariant(transportType));
+            }
+        }
+
+        static ITransportSettings[] PopulateCertificateInTransportSettings(IotHubConnectionStringBuilder connectionStringBuilder, ITransportSettings[] transportSettings)
+        {
+            foreach (var transportSetting in transportSettings)
+            {
+                switch (transportSetting.GetTransportType())
+                {
+                    case TransportType.Amqp_WebSocket_Only:
+                    case TransportType.Amqp_Tcp_Only:
+                        ((AmqpTransportSettings)transportSetting).ClientCertificate = connectionStringBuilder.Certificate;
+                        break;
+                    case TransportType.Http1:
+                        ((Http1TransportSettings)transportSetting).ClientCertificate = connectionStringBuilder.Certificate;
+                        break;
+                    case TransportType.Mqtt_WebSocket_Only:
+                    case TransportType.Mqtt_Tcp_Only:
+                        ((MqttTransportSettings)transportSetting).ClientCertificate = connectionStringBuilder.Certificate;
+                        break;
+                    default:
+                        throw new InvalidOperationException("Unsupported Transport {0}".FormatInvariant(transportSetting.GetTransportType()));
+                }
+            }
+
+            return transportSettings;
+        }
+    }
+}
